fix module resolution during formatting
