+++ cmd/templ/generatecmd/eventhandler.go
@@ -19,6 +19,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/a-h/templ/cmd/templ/generatecmd/modcheck"
 	"github.com/a-h/templ/cmd/templ/visualize"
 	"github.com/a-h/templ/generator"
 	"github.com/a-h/templ/parser/v2"
@@ -243,7 +244,15 @@ func (h *FSEventHandler) generate(ctx context.Context, fileName string) (result 
 	var b bytes.Buffer
 	// Use the directory of the specific file being processed for symbol resolution
 	fileDir := filepath.Dir(fileName)
-	generatorOutput, err := generator.Generate(t, &b, append(h.genOpts, generator.WithFileName(relFilePath), generator.WithWorkingDir(fileDir))...)
+
+	// Check if there's a go.mod file in fileDir or any parent directory up to that level
+	// If found, use that directory as working directory for proper module resolution
+	workingDir := fileDir
+	if modDir, err := modcheck.WalkUp(fileDir); err == nil {
+		workingDir = modDir
+	}
+
+	generatorOutput, err := generator.Generate(t, &b, append(h.genOpts, generator.WithFileName(relFilePath), generator.WithWorkingDir(workingDir))...)
 
 	// Always extract diagnostics from the generator, even if generation failed
 	parsedDiagnostics, diagErr := parser.Diagnose(t)
