+++ cmd/templ/imports/process.go
@@ -7,6 +7,7 @@ import (
 	"go/format"
 	"go/token"
 	"path"
+	"path/filepath"
 	"slices"
 	"strconv"
 	"strings"
@@ -17,6 +18,7 @@ import (
 	"golang.org/x/tools/go/ast/astutil"
 	"golang.org/x/tools/imports"
 
+	"github.com/a-h/templ/cmd/templ/generatecmd/modcheck"
 	"github.com/a-h/templ/generator"
 	"github.com/a-h/templ/parser/v2"
 )
@@ -79,7 +81,19 @@ func Process(t *parser.TemplateFile) (*parser.TemplateFile, error) {
 	var updatedImports []*ast.ImportSpec
 	var eg errgroup.Group
 	eg.Go(func() (err error) {
-		if _, err := generator.Generate(t, gw); err != nil {
+		// Determine the correct working directory for symbol resolution
+		var opts []generator.GenerateOpt
+		if t.Filepath != "" {
+			fileDir := filepath.Dir(t.Filepath)
+			// Check if there's a go.mod file in fileDir or any parent directory
+			if modDir, err := modcheck.WalkUp(fileDir); err == nil {
+				opts = append(opts, generator.WithWorkingDir(modDir))
+			} else {
+				opts = append(opts, generator.WithWorkingDir(fileDir))
+			}
+		}
+
+		if _, err := generator.Generate(t, gw, opts...); err != nil {
 			return fmt.Errorf("failed to generate go code: %w", err)
 		}
 		updatedImports, err = updateImports(fileName, gw.String())
