+++ generator/test-element-component/template.templ
@@ -3,8 +3,10 @@ package testelementcomponent
 import (
 	"context"
 	"fmt"
-	"github.com/a-h/templ/generator/test-element-component/mod"
 	"io"
+
+	"github.com/a-h/templ/generator/test-element-component/mod"
+	extern "github.com/a-h/templ/generator/test-element-component/externmod"
 )
 
 templ ElementComponent(attrs templ.Attributer) {
